deployment/requirements.txt
streamlit
pandas<2.0.0
seaborn
matplotlib
plotly
Pillow
altair
feature_engine
scikit-learn==1.2.1
tensorflow==2.12.0